2 / 15 / 2023
First Commit - Trivia Quiz App: question category array made containing an array holding the
text to be displayed to the user and the text that will be used for the api calling the
questions.

Messageboard component created using states for title and message.

CategoryCard component created, receives text, endpointtext, and a function to generate the
category section of the endpoint for calling the Trivia API on click.2