2 / 16 / 2023
Function added to change state when a category is selected, state determines if the component
should be displayed or not. Terniary operator checks if the chosen state is present before
removing the component from display.

DifficultyCard component created - Only displayed after category is chosen. Function created
to remove from the display the non-chosen difficulties. Grid Styling used to place the components
on the screen as sections are removed and added.