2 / 18 / 2023
Create AnswerCard.jsx to display the answers which are mapped to the cards via the
shuffled array of answers created in the useEffect Hook.

Create CheckAnswer function to check the selected card against the CorrectAnswer State.

Create returnToTitle function that resets the states managing the game progress to return
the user to the category selection

Create Score, totalScore, currentCategory, previousCategory states. Set up the CategoryCard
creation to check if category has been previously selected and to hide the card from display
if so.