2 / 21 /2023
Create GameOverModal to display when all categories have been chosen.
Style the GameOverModal to put shadow over the background and pop the modal on screen.

End of game proceedure created with a useEffect hook looking for when the length of the
chosen categories state has matched the list of categories [hard coded because no plans to change]

Bug found and corrected:
Bug - Total count was incorrect.
Issue - Was using the current score to increase total score rather than just increasing by one
Resolution - Changed the calculation to just increase the total score by 1.

Bug found and corrected:
Bug - Title Screen not displaying correct Total score
Issue - The setTitle() function in the returnToTitle() function was using the State for the
totalScore before it was updated with the final answer.
Solution - Created a useEffect Hook that will update the Title with the correct total score
value while the quizStart state is set to False.