2 / 22 / 2023
Alter color palet of page to appear softer with pastel colors.
Edit card CSS so they no longer change size on hover.
Bug Discovered - Category Cards not removing after the round has finished.
Bug Solution - The "btn" class was still applied to the buttons so removed this so that only
the "removed-card" class is on the removed cards.