2 / 23 / 2023
Set up states to check if player wishes to save their data then display a form for the user
to input their name. After submitting data, player will be returned to the game over screen
with the option to save removed and the message changed to show that the data has been saved.2

Next - Add a backend to the app to keep track of the data!